fix(shared): move service signature helper to a stdlib-only module (#12814)
